Refactor the project and the functions are verified to be working.
