+++ IntelliHealthApp/IntelliHealthSpring/src/main/java/com/intellihealthapp/main/IntelliHealthAppApplication.java
@@ -1,0 +1,13 @@
+package com.intellihealthapp.main;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class IntelliHealthAppApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(IntelliHealthAppApplication.class, args);
+	}
+
+}
