--- IntelliHealthApp/src/main/java/com/intellihealthapp/main/ServletInitializer.java
@@ -1,13 +1,0 @@
-package com.intellihealthapp.main;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(IntelliHealthAppApplication.class);
-	}
-
-}
