--- IntelliHealthApp/src/test/java/com/intellihealthapp/main/IntelliHealthAppApplicationTests.java
@@ -1,13 +1,0 @@
-package com.intellihealthapp.main;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class IntelliHealthAppApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
